More progress on getting threading working.
The datasource(s) is executed as a thread. Datasource is currently only
listserver. Datasource fetches data and queues them for a DataController. The
task of the DataController is to read these data and store them and possible
update QueryResult sets, so the GUI can be updated appropiately. The
implementation is still very experimental.